--- .travis.yml
@@ -1,8 +1,0 @@
-language: ruby
-bundler_args: --without development
-before_script: npm install -g bower
-rvm:
-  - 2.0.0
-branches:
-  except:
-    - front-end
