+++ spec/spec_helper.rb
@@ -1,0 +1,9 @@
+# Rails env
+require File.expand_path('../../config/environment', __FILE__)
+
+RSpec.configure do |config|
+  config.treat_symbols_as_metadata_keys_with_true_values = true
+  config.run_all_when_everything_filtered = true
+  config.filter_run :focus
+  config.order = 'random'
+end
