--- test/test_helper.rb
@@ -1,38 +1,0 @@
-ENV["RAILS_ENV"] = "test"
-require File.expand_path("../../config/environment", __FILE__)
-require "rails/test_help"
-require "minitest/rails"
-
-# To add Capybara feature tests add `gem "minitest-rails-capybara"`
-# to the test group in the Gemfile and uncomment the following:
-# require "minitest/rails/capybara"
-
-# Uncomment for awesome colorful output
-require "minitest/pride"
-
-class ActiveSupport::TestCase
-  # Setup all fixtures in test/fixtures/*.(yml|csv) for all tests in alphabetical order.
-  # fixtures :all
-
-  # Add more helper methods to be used by all tests here...
-end
-
-class Minitest::Test
-  def silence_stream(stream)
-    old_stream = stream.dup
-    stream.reopen('/dev/null')
-    stream.sync = true
-    yield
-  ensure
-    stream.reopen(old_stream)
-  end
-
-  def silence
-    silence_stream(STDERR) do
-      silence_stream(STDOUT) do
-        yield
-      end
-    end
-  end
-end
-
